requirements.txt
gradio>=4.0.0
numpy>=1.21.0
Pillow>=9.0.0
moviepy>=1.0.3
pytube>=15.0.0
TTS>=0.22.0
decord>=0.6.0
torch>=2.0.0
transformers>=4.30.0
langchain>=0.1.0
soundfile>=0.12.1
scipy>=1.10.0
scenedetect>=0.6.0
sentencepiece>=0.1.99
protobuf>=3.20.0
sacremoses>=0.0.53 